requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
docopt==0.4.0
gunicorn==19.1.1
mandrill==1.0.57
requests==2.4.1
static3==0.5.1
wsgiref==0.1.2
